Added te feature of "labels" to resources.
Now a user's resource can be accessed through the Stream API simply using its "label", i.e.
http://localhost:9000/streams/[username]/[label]

When in actuality, the label is pointing towards
http://localhost:9000/streams/[username]/[device]/[resource]
or 
http://localhost:9000/streams/[username]/[device]

The latter will only be useful is allow devices to present a resource interface by
default, avoiding the need to define a "/" resource for a website.

The original resource name is still accessible.

All of a users' active labels are now also displayed in their "Manage Resources" page,
and the set, display, boilerplate code should be in place behind the scenes. A new
subview directory: "views/vlabel" has been created to simplify display, though some
more clever object/inherritance is probably required for the future.

Please shout if anything breaks.